+++ Makefile.toml
@@ -11,7 +11,7 @@ extend = [
   { path = "src/risedevtool/risedev-components.toml" },
   { path = "src/frontend/planner_test/planner_test.toml" },
   { path = "src/tests/compaction_test/Makefile.toml" },
-  { path = "src/meta/src/backup_restore/integration_tests/Makefile.toml" },
+  { path = "src/storage/backup/integration_tests/Makefile.toml" },
 ]
 
 env_files = ["./risedev-components.user.env"]
