+++ src/cmd/Cargo.toml
@@ -16,6 +16,7 @@ anyhow = "1"
 clap = { version = "3", features = ["derive"] }
 global_stats_alloc = { path = "../utils/global_stats_alloc" }
 log = { version = "0.4" }
+risingwave_backup = { path = "../storage/backup" }
 risingwave_compaction_test = { path = "../tests/compaction_test" }
 risingwave_compactor = { path = "../storage/compactor" }
 risingwave_compute = { path = "../compute" }
