+++ src/meta/Cargo.toml
@@ -38,6 +38,7 @@ prometheus-http-query = "0.6"
 prost = "0.11"
 rand = "0.8"
 reqwest = "0.11"
+risingwave_backup = { path = "../storage/backup" }
 risingwave_common = { path = "../common" }
 risingwave_common_service = { path = "../common/common_service" }
 risingwave_connector = { path = "../connector" }
@@ -66,7 +67,6 @@ toml = "0.5"
 tonic = { version = "0.2", package = "madsim-tonic" }
 tower = { version = "0.4", features = ["util", "load-shed"] }
 tracing = "0.1"
-twox-hash = "1"
 url = "2"
 uuid = { version = "1", features = ["v4"] }
