+++ src/meta/src/lib.rs
@@ -34,6 +34,7 @@
 #![feature(let_chains)]
 #![feature(error_generic_member_access)]
 #![feature(provide_any)]
+#![feature(assert_matches)]
 #![cfg_attr(coverage, feature(no_coverage))]
 #![test_runner(risingwave_test_runner::test_runner::run_failpont_tests)]
