+++ src/storage/backup/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "risingwave_backup"
+version = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+keywords = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
+anyhow = "1"
+async-trait = "0.1"
+bytes = { version = "1", features = ["serde"] }
+itertools = "0.10"
+parking_lot = { version = "0.12", features = ["arc_lock"] }
+prost = "0.11"
+risingwave_common = { path = "../../common" }
+risingwave_hummock_sdk = { path = "../hummock_sdk" }
+risingwave_object_store = { path = "../../object_store" }
+risingwave_pb = { path = "../../prost" }
+serde = { version = "1", features = ["derive"] }
+serde_derive = "1"
+serde_json = "1"
+thiserror = "1"
+twox-hash = "1"
